変数展開エラー時のメモリリーク修正
