+++ include/parser/expansion.h
@@ -6,7 +6,7 @@
 /*   By: susumuyagi <susumuyagi@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/07 16:58:41 by susumuyagi        #+#    #+#             */
-/*   Updated: 2024/04/10 12:19:33 by susumuyagi       ###   ########.fr       */
+/*   Updated: 2024/04/11 17:51:53 by susumuyagi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,8 +30,8 @@ typedef struct s_expansion
 }					t_expansion;
 
 int					init_expansion(t_expansion *exp, t_token *tok);
+int					free_expansion_and_return_error(t_expansion *exp);
 int					expansion_done(t_expansion *exp);
-bool				is_special_param(t_expansion *exp);
 bool				has_quotes(t_token *tok);
 void				update_inside_status(t_expansion *exp);
 
@@ -42,6 +42,7 @@ int					join_var(t_expansion *exp, const char *var, size_t var_len);
 int					expand_variable(t_minishell *minish, t_expansion *exp);
 int					consume_and_join_dollar(t_expansion *exp);
 
+bool				is_special_param(t_expansion *exp);
 int					expand_special_param(t_minishell *minish, t_expansion *exp);
 
 char				*expand(t_minishell *minish, t_token *tok);
